Add CentOS support for role bird
