+++ molecule/delegated/tests/bird/main.py
@@ -1,4 +1,4 @@
-from .util.util import get_ansible, get_variable
+from ..util.util import get_ansible, get_variable
 
 testinfra_runner, testinfra_hosts = get_ansible()
 
@@ -9,19 +9,6 @@ def test_pkg(host):
 
     package = host.package(package_name)
     assert package.is_installed
-
-
-def test_conffile(host):
-    with host.sudo("bird"):
-        f = host.file("/etc/bird/bird.conf")
-        assert f.exists
-        assert not f.is_directory
-        assert f.user == "bird"
-        assert f.group == "bird"
-        assert f.mode == 0o640
-
-        bird_cidr = get_variable(host, "bird_cidr")
-        assert f"if net ~ [ {bird_cidr}+ ] then " in f.content_string
 
 
 def test_sysctl(host):
