+++ roles/bird/tasks/install-RedHat.yml
@@ -1,0 +1,46 @@
+---
+- name: Install EPEL repository
+  become: true
+  ansible.builtin.dnf:
+    name: epel-release
+    state: present
+  changed_when: true
+
+- name: Enable EPEL repository
+  become: true
+  ansible.builtin.command:
+    cmd: dnf config-manager --set-enabled epel
+  changed_when: true
+
+- name: Update package cache after installing EPEL repo
+  become: true
+  ansible.builtin.dnf:
+    update_cache: true
+  changed_when: true
+
+- name: Install bird package
+  become: true
+  ansible.builtin.dnf:
+    name: "{{ bird_package_name }}"
+    state: present
+
+#  Config file needs to be adapted for bird 2.x
+# - name: Copy bird.conf configuration file
+#   become: true
+#   ansible.builtin.template:
+#     src: bird.conf.j2
+#     dest: /etc/bird.conf
+#     owner: bird
+#     group: bird
+#     mode: 0640
+#   notify: Restart bird service
+
+# This can be removed, if the task above is fixed:
+- name: Copy bird.conf configuration file
+  become: true
+  ansible.builtin.file:
+    path: /etc/bird.conf
+    owner: bird
+    group: bird
+    mode: 0640
+  notify: Restart bird service
